Conflicts:
	C#/BDSAE2013/BDSAE2013.sln

+++ C#/BDSAE2013/Storage/StorageConnectionBridgeFacade.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 
 namespace Storage
@@ -6,8 +7,22 @@ namespace Storage
     /// <summary>
     /// Refined IStorageBridge implementation.
     /// </summary>
+    /// <author>
+    /// Jacob Cholewa (jbec@itu.dk)
+    /// </author>
     public class StorageConnectionBridgeFacade : AbstractStorageConnectionBridgeFacade
     {
+        private static readonly Dictionary<Type,Integer> IdCounts = new Dictionary<Type, Integer>();
+
+        private static Integer GetId<TEntity>()
+        {
+            if (!IdCounts.ContainsKey(typeof (TEntity)))
+            {
+                IdCounts[typeof(TEntity)] = new Integer();
+            }
+            return IdCounts[typeof (TEntity)];
+        }
+
         /// <summary>
         /// Constructs a new StorageBridgeFacade
         /// </summary>
@@ -73,29 +88,20 @@ namespace Storage
             IsDisposed();
 
             //Makes sure the entity is not null
-            if(entity == null) throw new ArgumentNullException("entity");
+            if (entity == null) throw new ArgumentNullException("entity");
 
             //Makes sure the entities id is not preset
-            if(entity.Id != 0) throw new ArgumentException("Id can't be preset!");
-
-            //Setting the id of the entity
-            try
-            {
-                entity.Id = Get<TEntity>().Max(t => t.Id) + 1;
-            }
-            catch (InvalidOperationException)
-            {
-                entity.Id = 1;
-            }
-
-            //Checks that the id has been set
-            if(entity.Id == 0) throw new InternalDbException("Id was not set");
+            if (entity.Id != 0) throw new ArgumentException("Id can't be preset!");
 
             //Adds the entity to the context
             Db.Add(entity);
 
             //Saves the context
             SaveChanges();
+
+            //Checks that the id has been set
+            if (entity.Id == 0) throw new InternalDbException("Id was not set");
+
         }
 
         /// <summary>
@@ -113,6 +119,7 @@ namespace Storage
         {
             IsDisposed(); //Checks that the context is not disposed
             if (entity == null){ throw new ArgumentNullException("entity");} //Checks that the entity is not null
+            if (entity.Id == 0) throw new InternalDbException("Id was not set");
             Db.Update(entity); //Updates the entity
             SaveChanges(); //Saves the changes to the context
         }
@@ -131,6 +138,7 @@ namespace Storage
         {
             IsDisposed(); //Checks that the context is not disposed
             if (entity == null){ throw new ArgumentNullException("entity"); } //Checks that the entity is not null
+            if (entity.Id == 0) throw new InternalDbException("Id was not set");
             Db.Delete(entity); //Deletes the entity
             SaveChanges(); //Saves the changes to the database
         }
@@ -151,6 +159,19 @@ namespace Storage
             if (id <= 0) { throw new ArgumentException("Ids 0 or less"); } //Checks that the id is not 0 or less
             if (id > int.MaxValue) { throw new ArgumentException("Ids larger than int.MaxValue"); } //Checks that the id is not more that int.maxValue
             Delete(Get<TEntity>(id)); //Calls Delete with entity fetched by id
+        }
+
+        /// <summary>
+        /// Helper class to solve concurrency when distributing ID's
+        /// </summary>
+        private class Integer
+        {
+            internal Integer()
+            {
+                Value = 0;
+            }
+
+            internal int Value { get; set; }
         }
     }
 }
